+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.26.0](https://github.com/TheRealKoller/photosort/compare/v0.25.0...v0.26.0) (2026-08-30)
+
+
+### Features
+
+* **scripts:** PR verknuepft sein Issue per Closing-Zeile, finalize verifiziert das (Spec 0251) ([#280](https://github.com/TheRealKoller/photosort/issues/280)) ([004ad48](https://github.com/TheRealKoller/photosort/commit/004ad48e6a6818aecbd29bd318232b2a966ed97f))
+
+## [0.25.0](https://github.com/TheRealKoller/photosort/compare/v0.24.0...v0.25.0) (2026-08-30)
+
+
+### Features
+
+* **frontend:** Bewertungsdetails in Bloecke Qualitaet und Kategorien gliedern (Spec 0209) ([#277](https://github.com/TheRealKoller/photosort/issues/277)) ([c673173](https://github.com/TheRealKoller/photosort/commit/c6731739ce5290ba263565a9a0ddc265def50a44))
+
 ## [0.24.0](https://github.com/TheRealKoller/photosort/compare/v0.23.0...v0.24.0) (2026-08-30)